+++ gazebo/physics/Joint.cc
@@ -35,6 +35,8 @@
 #include "gazebo/physics/World.hh"
 #include "gazebo/physics/Joint.hh"
 
+#include "gazebo/util/IntrospectionManager.hh"
+
 using namespace gazebo;
 using namespace physics;
 
@@ -1447,4 +1449,47 @@ math::Pose Joint::ComputeChildLinkPose(unsigned int _index,
   }
 
   return newWorldPose;
+}
+
+/////////////////////////////////////////////////
+void Joint::RegisterIntrospectionItems()
+{
+  for (size_t i = 0; i < this->GetAngleCount(); ++i)
+  {
+    this->RegisterIntrospectionPosition(i);
+    this->RegisterIntrospectionVelocity(i);
+  }
+}
+
+/////////////////////////////////////////////////
+void Joint::RegisterIntrospectionPosition(const unsigned int _index)
+{
+  auto f = [this, _index]()
+  {
+    // For prismatic axes, Radian -> meters
+    return this->GetAngle(_index).Ign().Radian();
+  };
+
+  common::URI uri(this->URI());
+  uri.Query().Insert("p",
+      "axis/" + std::to_string(_index) + "/double/position");
+  this->introspectionItems.push_back(uri);
+  gazebo::util::IntrospectionManager::Instance()->Register
+      <double>(uri.Str(), f);
+}
+
+/////////////////////////////////////////////////
+void Joint::RegisterIntrospectionVelocity(const unsigned int _index)
+{
+  auto f = [this, _index]()
+  {
+    return this->GetVelocity(_index);
+  };
+
+  common::URI uri(this->URI());
+  uri.Query().Insert("p",
+      "axis/" + std::to_string(_index) + "/double/velocity");
+  this->introspectionItems.push_back(uri);
+  gazebo::util::IntrospectionManager::Instance()->Register
+      <double>(uri.Str(), f);
 }